+++ android/src/main/kotlin/com/hisaichi5518/native_webview/InputAwareWebView.java
@@ -6,14 +6,11 @@ package com.hisaichi5518.native_webview;
 
 import static android.content.Context.INPUT_METHOD_SERVICE;
 
-import android.app.Activity;
 import android.content.Context;
 import android.graphics.Rect;
 import android.os.Build;
-import android.os.IBinder;
 import android.util.Log;
 import android.view.View;
-import android.view.Window;
 import android.view.inputmethod.InputMethodManager;
 import android.webkit.WebView;
 import android.widget.ListPopupWindow;