+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix [Keyboard persists after tapping outside text field](https://github.com/flutter/flutter/issues/36478#issuecomment-623542124)
 - Fix [The keyboard is not displayed in the example application.](https://github.com/hisaichi5518/native_webview/pull/45)
+- [Display selection popup on Android WebView](https://github.com/hisaichi5518/native_webview/pull/48)
 
 ## 0.15.0